requirements.txt
alembic==1.9.4
aniso8601==9.0.1
Bootstrap-Flask==2.2.0
CacheControl==0.13.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==41.0.1
distlib==0.3.6
filelock==3.12.0
firebase-admin==6.1.0
fireo==2.1.0
Flask==2.2.2
Flask-Migrate==4.0.4
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.3
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.54.2
grpcio-status==1.54.2
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
msgpack==1.0.5
pika==1.3.1
pika-stubs==0.1.3
platformdirs==3.5.1
proto-plus==1.22.2
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.0.9
python-dotenv==0.21.1
pytz==2023.3
PyYAML==6.0
requests==2.28.2
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.23.0
Werkzeug==2.2.2
WTForms==3.0.1
